refactor lib.rs to linux.rs
